PHPUnit tests + CodeSniffer
