+++ .travel.yml
@@ -1,0 +1,14 @@
+language: php
+
+php:
+  - 5.6
+  - 7.0
+  - hhvm
+
+before_script:
+  - composer self-update
+  - composer install --prefer-source --no-interaction --dev
+
+script:
+  - vendor/bin/phpunit
+  - vendor/bin/phpcs src --standard=PSR2
