+++ tests/User.php
@@ -1,0 +1,6 @@
+<?php
+
+class User extends \App\User
+{
+    use \Inani\Messager\Helpers\MessageAccessible;
+}
